Added Connected button and update the status of the button if the candidate is connected as a friend.
add a remove friend button so that you can unfollow the candidate.

once the connect button is clicked updated the firebase by adding an extra field in the Candidates as "friends"

displayed the connected friends in the main activity in the friend list in a recycler view.

updated some design files